Move RecordingDescriptor into SubprocessReportGenerator
